+++ demisto_sdk/commands/format/update_incidenttype.py
@@ -27,14 +27,15 @@ ARGUMENTS_TO_REMOVE = ['sortValues',
 class IncidentTypesJSONFormat(BaseUpdateJSON):
     """IncidentTypesJSONFormat class is designed to update incident fields JSON file according to Demisto's convention.
 
+
         Attributes:
-            source_file (str): the path to the file we are updating at the moment.
-            output_file_name (str): the desired file name to save the updated version of the YML to.
+            input (str): the path to the file we are updating at the moment.
+            output (str): the desired file name to save the updated version of the YML to.
             json_data (Dict): YML file data arranged in a Dict.
     """
 
-    def __init__(self, source_file='', output_file_name='', old_file=''):
-        super().__init__(source_file, output_file_name, old_file)
+    def __init__(self, input='', output='', old_file=''):
+        super().__init__(input, output, old_file)
 
     def format_file(self):
         """Manager function for the integration YML updater."""