+++ demisto_sdk/commands/format/update_integration.py
@@ -10,8 +10,8 @@ class IntegrationYMLFormat(BaseUpdateYML):
     """PlaybookYMLFormat class is designed to update integration YML file according to Demisto's convention.
 
         Attributes:
-            source_file (str): the path to the file we are updating at the moment.
-            output_file_name (str): the desired file name to save the updated version of the YML to.
+            input (str): the path to the file we are updating at the moment.
+            output (str): the desired file name to save the updated version of the YML to.
             yml_data (Dict): YML file data arranged in a Dict.
             id_and_version_location (Dict): the object in the yml_data that holds the is and version values.
     """
@@ -21,8 +21,8 @@ class IntegrationYMLFormat(BaseUpdateYML):
         'proxy': 'Use system proxy settings'
     }
 
-    def __init__(self, source_file='', output_file_name='', old_file=''):
-        super().__init__(source_file, output_file_name, old_file)
+    def __init__(self, input='', output='', old_file=''):
+        super().__init__(input, output, old_file)
 
     def update_proxy_insecure_param_to_default(self):
         """Updates important integration arguments names and description."""