+++ demisto_sdk/commands/format/update_script.py
@@ -7,14 +7,14 @@ class ScriptYMLFormat(BaseUpdateYML):
     """ScriptYMLFormat class is designed to update script YML file according to Demisto's convention.
 
         Attributes:
-            source_file (str): the path to the file we are updating at the moment.
-            output_file_name (str): the desired file name to save the updated version of the YML to.
+            input (str): the path to the file we are updating at the moment.
+            output (str): the desired file name to save the updated version of the YML to.
             yml_data (Dict): YML file data arranged in a Dict.
             id_and_version_location (Dict): the object in the yml_data that holds the is and version values.
     """
 
-    def __init__(self, source_file='', output_file_name='', old_file=''):
-        super().__init__(source_file, output_file_name, old_file)
+    def __init__(self, input='', output='', old_file=''):
+        super().__init__(input, output, old_file)
 
     def format_file(self):
         """Manager function for the script YML updater."""